Add Link To Readme
I added the updated link to Nexum. For some reason, trying to access the main page results in a 404 Error. I'll try to look into it when I get the chance and keep y'all posted. It would be wonderful if a second pair of eyes could try to debug the error as well, as I don't have any leads thus far.